% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractMultiDataSet.R
\docType{data}
\name{AbstractMultiDataSet}
\alias{AbstractMultiDataSet}
\alias{AbstractMultiDataSet}
\title{An S6 class representing a generic collection of datasets}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbstractMultiDataSet
}
\arguments{
\item{key1}{Numeric or character index of first dataset to use}

\item{key2}{Numeric or character index of second dataset to use}

\item{method}{Correlation method to use (passed to \code{cor} function)}

\item{key1}{Numeric or character index of first dataset to use}

\item{key2}{Numeric or character index of second dataset to use}

\item{method}{Correlation method to use (passed to \code{cor} function)

Creates a static or interactive heatmap plot}

\item{params}{A list of plotting parameters}

\item{interactive}{Logical indicating whether an interactive heatmap
should be generated.}
}
\value{
Matrix of pairwise dataset1 - dataset2 correlations
Plots multidataset correlation heatmap
}
\description{
An S6 class representing a generic collection of datasets

private params
Computes cross-dataset correlation matrix
}
\keyword{datasets}
