% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_numeric_logic.R
\name{apply_numeric_logic}
\alias{apply_numeric_logic}
\title{apply_numeric_logic}
\usage{
apply_numeric_logic(main_df, logic_df, unique_key_id, logic_type = "woe")
}
\arguments{
\item{main_df}{A dataframe that you want to transform.  Ensure that it has a unique field which will be represented in the unique_key_id}

\item{logic_df}{A dataframe which provides the logic to apply to each variable}

\item{unique_key_id}{A single attribute representing the unique key in the dataframe}

\item{logic_type}{The logic to use to create either woe and/or grp variables. Options are c('woe','grp')}
}
\value{
A dataframe withe the unique_key_id field along with all transformed attributes (woe and/or grp).  The original attributes will not be in this data frame.  The only original variable that will be in this data frame is the unique_key_id field
}
\description{
Apply numeric grouping logic
}
