% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicSummary.R
\name{basicSummary}
\alias{basicSummary}
\title{Basic Summary}
\usage{
basicSummary(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
NULL.
}
\description{
This function receives as arguments the target dataframe. It returns a very
basic descriptive of the dataframe: the number of observations, the number
of variables, how many numeric, categoric and factor variables there are and the
percentage of complete cases (rows without NAs) and complete variables
(columns without NA).
}
