% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioExprSet.R
\name{BioExprSet}
\alias{BioExprSet}
\title{An R6 class representing a high-throughput expression dataset along with
associated gene and sample metadata.}
\description{
An R6 class representing a high-throughput expression dataset along with
associated gene and sample metadata.
}
\section{Usage}{
\preformatted{# TODO
}
}

\section{Arguments}{

\itemize{
\item \code{datasets}: A list of datasets (matrices, data frames, etc.), each of
which shared some column / row identifiers with the first entry in
the list.
}
}

\section{Fields}{

\itemize{
\item \code{datasets}: List of datasets
}
}

\section{Methods}{

\itemize{
\item \code{cross_cor(key1=1, key2=2, meas='pearson')}: Computes cross-dataset
correlation matrix between rows in two specified datasets.
\item \code{plot_cross_cor_heatmap(key1=1, key2=2, meas='pearson', interactive=TRUE)}:
Plots multidataset correlation heatmap.
\item \code{print()}: Prints an overview of the object instance.
}
}

\section{Examples}{
\preformatted{TODO
}
}

