% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarbonEmission}
\alias{CarbonEmission}
\alias{carbon}
\alias{energy}
\alias{buildingarea}
\alias{factordata}
\title{Carbon emission data for energy decomposition analysis.}
\format{\code{carbon}: A data frame of carbon emissions from multiple types 
of fuels and building sectors.
\code{energy}: A data frame energy consumptions from multiple types 
of fuels and building sectors.
\code{buildingarea}: A data frame of annual building areas.
\code{factordata}: A list of data frames of factors of carbon emission changes.}
\usage{
carbon

energy

buildingarea

factordata
}
\description{
Carbon emission data for energy decomposition analysis.
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{Carbon}
\keyword{Emission}
\keyword{datasets}
