% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_dropdown.R
\name{contact_item}
\alias{contact_item}
\title{Contact Item}
\usage{
contact_item(
  name = "First Name, Last Name",
  role = "Role",
  phone = "###-###-####",
  email = "first.last@oliverwyman.com"
)
}
\arguments{
\item{name}{Name}

\item{role}{Role}

\item{phone}{Phone}

\item{email}{Email}
}
\value{
contact menu item
}
\description{
Creates an item to be placed in a contact dropdownmenu.
}
