% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_dropdown.R
\name{contact_menu}
\alias{contact_menu}
\title{Creates a dropdown menu specific for contacts}
\usage{
contact_menu(...)
}
\arguments{
\item{...}{contact items to put into dropdown}
}
\value{
menu
}
\description{
Creates a dropdown menu specific for contacts
}
