% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_classes.R
\name{count_classes}
\alias{count_classes}
\title{Title}
\usage{
count_classes(data_frame, class_col)
}
\arguments{
\item{class_col}{}
}
\value{
#'   The first column (named class) lists the classes from the input data frame.
The second column (named count) lists the number of observations for each class from the input data frame.
It will have one row for each class present in input data frame.
}
\description{
Title
}
