% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{create_analysis_table}
\alias{create_analysis_table}
\alias{create_numeric_table}
\title{Create EDA table from dataset}
\usage{
create_analysis_table(df, var_name, target_var, type = "count")

create_numeric_table(df, var_name, target_var, type = "count",
  bins = 30)
}
\arguments{
\item{df}{A \code{data.frame} with the dataset to analyze}

\item{var_name}{A string with the variable name to analyze}

\item{target_var}{A string with the name of the target variable}

\item{type}{Can be \code{'count'} or \code{'average'} depicting how to get
the results back}

\item{bins}{Number of bins to use for numeric variables}
}
\description{
These functions create tables summarizing how the variable \code{var_name} in
the dataset \code{df} relates to the specified target variable
\code{target_var}.
}
\details{
You can specify the \code{type} as \code{'count'} or \code{'average'}. Use
\code{'count'} to get counts of each variable value for the values of the
target variable. Use \code{'average'} to get the average of the target
variable for variable value.

In the case of \code{create_numeric_table} the input variable is a numeric
variable and its values are created as bins, with number of bins defined as
\code{bins}
}
