% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{draw_boxplot}
\alias{draw_boxplot}
\title{Title}
\usage{
draw_boxplot(
  wb,
  x_var,
  y_var,
  type = "Boxplot",
  p_format = "p.adj",
  porownanie = "brak",
  punkty = "none",
  anova = "nie",
  test_type = "t.test",
  kontrola = 1,
  grupy_porownania = "",
  x_name = "x",
  y_name = "y",
  kolory = "default",
  viridis = "magma",
  brewer = "Set1",
  wlasne
)
}
\arguments{
\item{wb}{dataframe}

\item{x_var}{variable for x axis}

\item{y_var}{variable for y axis}

\item{type}{plot type}

\item{p_format}{format for p value}

\item{porownanie}{comparison type}

\item{punkty}{draw points? based upon ggbeeswarm package}

\item{anova}{show anova result?}

\item{test_type}{which statistical test should be used?}

\item{grupy_porownania}{which groups should be compared}

\item{x_name}{x axis title}

\item{y_name}{y axis title}

\item{kolory}{color scale type}

\item{viridis}{viridis scale type}

\item{brewer}{colorbrewer palette}

\item{wlasne}{custom colors}

\item{konrola}{which group is the control group?}
}
\value{
ggplot
}
\description{
Draws boxplot, violin of mean_ci plot based upon ggpubr package
}
