% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{draw_density}
\alias{draw_density}
\title{Title}
\usage{
draw_density(
  wb,
  variable,
  color_var,
  fill = FALSE,
  x_name = "x",
  y_name = "y",
  kolory = "default",
  viridis = "magma",
  brewer = "Set1",
  wlasne
)
}
\arguments{
\item{wb}{dataframe}

\item{variable}{name for x}

\item{color_var}{name for facet/colorscale}

\item{fill}{add fill to density curves?}

\item{x_name}{x label}

\item{y_name}{y label}

\item{kolory}{choose colour scale}

\item{viridis}{choose viridis scale}

\item{brewer}{choose colorbrewer scale}

\item{wlasne}{choose custom scale}
}
\value{
ggplot
}
\description{
Title
}
