% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{draw_histogram}
\alias{draw_histogram}
\title{Title}
\usage{
draw_histogram(
  wb,
  variable,
  facet_draw = TRUE,
  facet_var,
  bin = 1,
  y_density = 1,
  x_name = "x",
  y_name = "y",
  kolory = "default",
  viridis = "magma",
  brewer = "Set1",
  wlasne
)
}
\arguments{
\item{wb}{dataframe}

\item{variable}{name for x}

\item{facet_draw}{draw facets (logical)}

\item{facet_var}{name for facet/colorscale}

\item{bin}{bin width}

\item{y_density}{density on y scale?}

\item{x_name}{x label}

\item{y_name}{y label}

\item{kolory}{choose colour scale}

\item{viridis}{choose viridis scale}

\item{brewer}{choose colorbrewer scale}

\item{wlasne}{choose custom scale}
}
\value{
ggplot
}
\description{
Draws a ggplot histogram
}
