% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{draw_scatter}
\alias{draw_scatter}
\title{Title}
\usage{
draw_scatter(
  wb,
  x_var,
  y_var,
  color_var = "none",
  facet_var = "none",
  trend = FALSE,
  size_trend = 2,
  model = "loess",
  span = 0.75,
  se = TRUE,
  alpha = 1,
  size_point = 1.5,
  kolory = "default",
  viridis = "magma",
  brewer = "Set1",
  wlasne
)
}
\arguments{
\item{wb}{data frame}

\item{x_var}{x axis variable}

\item{y_var}{y axis variable}

\item{color_var}{color variable}

\item{facet_var}{facet_wrap variable}

\item{trend}{add trendline?}

\item{size_trend}{size of trend line}

\item{model}{model, option: 'loess', 'lm'}

\item{span}{span of loess}

\item{se}{add error?}

\item{alpha}{alpha of points}

\item{size_point}{size of points}

\item{kolory}{color type}

\item{viridis}{viridis scale}

\item{brewer}{colorbrewer scale}

\item{wlasne}{custom colors}
}
\value{
ggplot
}
\description{
Title
}
