% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA.R
\name{EDA}
\alias{EDA}
\title{An R6 class representing one or more datasets linked by either column or row identifiers.}
\description{
An R6 class representing one or more datasets linked by either column or row identifiers.
}
