% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDADat.R
\name{EDADat}
\alias{EDADat}
\title{An R6 class representing a single dataset.}
\description{
EDADat is a container class representing a single dataset (matrix or data
frame) along with some basic information about its indexing, orientation,
and representation.
}
\details{
The purpose of this class is to provide a way for each atomic dataset to
carry its own relevant metadata, as needed by the other major \code{eda} classes.
This class is different from all other \code{eda} classes in that it does not
inherit from \code{AbstractMultiDataSet}, and, aside from being passed into other
\code{eda} class constructors, it is not intended to be used directly by the user.
}
\section{Arguments}{

\itemize{
\item \code{dat}: A data frame or matrix
\item \code{key}: Character string or number indicating the row or column containg
the dataset primary keys. If 'rownames', or 'colnames', row or column
names will be used, respectively. If NULL, defaults to 'rownames' or
'colnames', depending on orientation.
\item \code{transposed}: Logical indicating whether the dataset orientation is
transposed, relative to the primary dataset. For example, if a primary
and secondary dataset both share the same column identifiers, but the
second dataset is oriented with those id's as rows, then transposed
should be set to TRUE to indicate such.
\code{xlab}: Character string containing x-axis label
\code{ylab}: Character string containing y-axis label
}
}

\section{Fields}{

\itemize{
\item \code{dat}: Formatted data frame or matrix
}
}

