% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDADataSet.R, R/EDAMultiDataSet.R
\docType{data}
\name{EDADataSet}
\alias{EDADataSet}
\alias{EDADataSet}
\alias{EDADataSet}
\title{An S6 class representing a generic dataset.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EDADataSet
}
\arguments{
\item{mask}{Logical vector of length equal to the number of rows in
the dataset.}

\item{mask}{Logical vector of length equal to the number of columns 
in the dataset.}

\item{method}{Character array specifying imputation method to use 
    (knn)
Summarizes overall characteristics of a dataset

Kernel density plot

Plots densities for each row or column in the dataset. This is most
useful when you are interested in similarties or differences in
distributions across columns, for a relatively small number of
columns.}

\item{color}{Variable to color density curves by. If not is
specified, uses variable specified at object construction time,
or else, uses a separate color for each column.}

\item{ind}{Numeric vector of row indices to use for intensive 
    operations when subsampling is enabled.
Sets the subsampling column indices for a dataset}

\item{ind}{Numeric vector of column indices to use for intensive 
    operations when subsampling is enabled.
Computes correlations between axes of a data projection (PCA, t-SNE,
etc.) and column metadata.}

\item{mat}{Numeric projected data matrix}

\item{include}{Vector of strings indicating metadata columns which
should be included in the analysis.}

\item{...}{Arguments passed to a given plotting, etc. function call.}

\item{...}{Arguments passed to a given plotting, etc. function call.}

\item{color}{Color variable as passed into plot function call}

\item{color}{Color variable as passed into plot function call}

\item{color}{Color variable as passed into plot function call}

\item{shape}{Shape variable as passed into plot function call}

\item{styles}{List of color-related style info}

\item{color}{Color variable passed into plot function call
Determines shape-related style information to use for a plot}

\item{styles}{List of shape-related style info}

\item{shape}{shape variable passed into plot function call
Returns a list of color-related style information}

\item{color}{Color variable as passed into plot function call}

\item{shape}{Shape variable as passed into plot function call}

\item{color}{Variable to use for assigning colors.}

\item{color_pal}{Color palette to use}

\item{shape}{Variable to use for assigning shapes.}

\item{params}{A list of plotting parameters}

\item{interactive}{Logical indicating whether an interactive heatmap
    should be generated.
Normalizes handling of data row and column identifiers

Checks dataset row and column identifiers and converts them to row 
and column names, respectively, if they are not already stored there.}

\item{dat}{Dataset}

\item{row_ids}{Column id or number where row identifiers are stored.}

\item{col_ids}{Row id or number where column identifiers are stored.}

\item{Dataset}{with identifiers as rows and columns
Prints dataset summary to screen
Determines metadata columns to include for heatmap, etc. functions}

\item{include}{Vector of strings indicating metadata columns which
should be included in the analysis.}

\item{exclude}{Vector of strings indicating metadata columns which
should be excluded from the analysis.}
}
\value{
A filtered version of the original EDADataSet object.
Applies a filter to columns of the dataset

A filtered version of the original EDADataSet object.
Imputes missing values in the dataset

Imputes missing values in the dataset using a specified method
and stores the result in-place. Currently only support k-Nearest
Neighbors (kNN) method.

Note: When using the `knn` method, it may be neccessary to set R 
the environmental variable `R_MAX_NUM_DLLS` to some value larger
than its default of 100 (150 should be sufficient), prior to
launching R. This can be set in the ~/.Renviron.

ggplot plot instance.

Prints class greeting to the screen
Gets the subsampling row indices for a dataset

Numeric vector of row indices to use for intensive 
    operations when subsampling is enabled.
Gets the subsampling column indices for a dataset

Numeric vector of column indices to use for intensive 
    operations when subsampling is enabled.
Sets the subsampling row indices for a dataset

Dataframe of feature x project axes dependencies
Returns a list of arguments specific for a given function call.

This function takes a list of function arguments and generates
a new list without any shared arguments (row_maxn, etc.).

A list containing only function-specific arguments 
Determines row and column indices to use for a function call

Parses a list of function arguments and check whether user has
specified any subsampling-related arguments (row_maxn, row_max_ratio,
etc.) and, if so, select indices to use. Otherwise object-level
indices are returned.

A list with 'row' and 'column' entries corresponding to the
    specific numeric row and column indices to be used.
Generates ggplot aesthetics for density plots

List of style information
Generates ggplot aesthetics for histogram plots

List of style information
Generates ggplot aesthetics for a geom_point plot

List of geom_point style information
Determines color-related style information to use for a plot

list List of color-related properties
Returns a list of shape-related style information

list List of shape-related properties
Returns a vector of color codes associated with the specified
variable.

Vector of colors with length equal to the number of columns
	       in the data.
Returns a vector of text labels to use for a plot

Vector of labels with length equal to the number of columns
        in the data.
Creates a static or interactive heatmap plot

Character vector of fields to include in analysis.
Transpose the dataset and metadata in-place
}
\description{
An S6 class representing a generic dataset.

Clears any cached resuts and performs garbage collection to free 
up memory.
Applies a filter to rows of the dataset

An S6 class representing collection of related datasets
}
\section{Arguments}{

\describe{
  \item{dat}{An m x n dataset.}
  \item{row_mdata}{Data frame with rows corresponding to the column names of 
      \code{dat}.}
  \item{row_mdata}{Data frame with rows corresponding to the row names of 
      \code{dat}.}
}
}

\keyword{datasets}
