% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAMatrix.R
\docType{data}
\name{EDAMatrix}
\alias{EDAMatrix}
\alias{EDAMatrix}
\title{An S6 class representing a matrix dataset.}
\format{\code{\link{R6Class}} object.}
\usage{
EDAMatrix
}
\value{
Vector of cluster assignments with length equal to the
number of rows in the dataset.
Detects column outliers in the dataset

Computes pairwise correlations between all columns in the dataset.
Columns whose median pairwise correlation with all other columns
is greater than \code{num_sd} standard deviations from the average
median correlation are considered to be outliers.

num_sd Number of standard deviations to use to determine
outliers.

Character vector or column ids for columns with low
average pairwise correlations.
Detects row outliers in the dataset

Computes pairwise correlations between all rows in the dataset.
Rows whose median pairwise correlation with all other rows
is greater than \code{num_sd} standard deviations from the average
median correlation are considered to be outliers.

num_sd Number of standard deviations to use to determine
outliers.

Character vector or row ids for rows with low
average pairwise correlations.
Detects dependencies between column metadata entries (features) and
dataset rows.
Removes column outliers from the dataset

Computes pairwise correlations between all columns in the dataset.
Columns whose median pairwise correlation with all other columns
is greater than \code{num_sd} standard deviations from the average
median correlation are removed.

num_sd Number of standard deviations to use to determine
outliers.

A filtered version of the original EDADataSet object.
Removes row outliers from the dataset

Computes pairwise correlations between all rows in the dataset.
rows whose median pairwise correlation with all other rows
is greater than \code{num_sd} standard deviations from the average
median correlation are removed.

num_sd Number of standard deviations to use to determine
outliers.

A filtered version of the original EDADataSet object.
Log-transforms data

base Numeric logarithm base to use (default: e)
offset Numeric offset to apply to data before taking the
logarithm (default: 0)

A log-transformed version of the object.
Log(x + 1) transforms data

A log(x + 1) transformed version of the object.
PCA

t-SNE

Prints an overview of the object instance
Correlation heatmap.

Generates a correlation heatmap depicting the pairwise column
correlations in the data.

method String name of correlation method to use.
... Additional arguments

ggplot plot instance
Generates a two-dimensional PCA plot from the dataset

pcx integer PC number to plot along x-axis (default: 1)
pcy integer PC number to plot along x-axis (default: 2)
scale Whether or not to scale variables prior to performing
pca; passed to \code{prcomp} function.
color Column metadata field to use for coloring points.
shape Column metadata field to use to assign shapes to points
title Plot title.
text_labels Whether or not to include individual point labels
plot (default: FALSE).
...

ggplot plot instance
Generates a two-dimensional t-SNE plot from the dataset

color Column metadata field to use for coloring points.
shape Column metadata field to use to assign shapes to points
title Plot title.
text_labels Whether or not to include individual point labels
plot (default: FALSE).
...

ggplot plot instance
Plot median pairwise column correlations

Plots the median correlation of each variable (column). This is
useful for visually inspecting columns for possible outliers, when
the total number of columns is relatively small.

None
Verifies that input data is of type matrix
}
\description{
EDAMatrix is a helper class for wrapping data matrices, with optional
support for row and column datadata. Methods are provided for common
exploratory data analysis summary statistics, transformations, and
visualizations.

k Number of clusters to detect (default: 10)
}
\seealso{
\code{cor} for more information about supported correlation
methods.
Generates a heatmap plot of the dataset

... Additional arguments
Creates a tile plot of projected data / feature correlations

include Vector of strings indicating metadata columns which
should be included in the analysis.
exclude Features (column metadata variables) to exclude from
the analysis.
color_scale Character vector containing colors to sue for
low-correlation and high-correlation values
(default: c('green', 'red')).
}
\author{
V. Keith Hughitt, \email{keith.hughitt@nih.gov}
}
\keyword{datasets}
