% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAMultiDataSet.R
\docType{data}
\name{EDAMultiDataSet}
\alias{EDAMultiDataSet}
\alias{EDAMultiDataSet}
\title{An S6 class representing collection of related datasets}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EDAMultiDataSet
}
\arguments{
\item{key1}{Numeric or character index of first dataset to use}

\item{key2}{Numeric or character index of second dataset to use}

\item{method}{Correlation method to use (passed to \code{cor} function)}

\item{key1}{Numeric or character index of first dataset to use}

\item{key2}{Numeric or character index of second dataset to use}

\item{method}{Correlation method to use (passed to \code{cor} function)

Prints an overview of the object instance
Make sure input datasets are all EDADataSet instances and include
the same column identifiers
Make datasets publically visible for EDAMultiDataSet instances}
}
\value{
Matrix of pairwise dataset1 - dataset2 correlations
Plots multidataset correlation heatmap
}
\description{
An S6 class representing collection of related datasets

Computes cross-dataset correlation matrix
}
\keyword{datasets}
