% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAMultiMatrix.R
\name{EDAMultiMatrix}
\alias{EDAMultiMatrix}
\title{An R6 class representing collection of related datasets with a matrix
for the primary dataset.}
\description{
An R6 class representing collection of related datasets with a matrix
for the primary dataset.
}
\section{Usage}{
\preformatted{# TODO
}
}

\section{Arguments}{

\itemize{
\item \code{dataset}: A list of datasets (matrices, data frames, etc.), each of
which shared some column / row identifiers with the first entry in
the list.
}
}

\section{Fields}{

\itemize{
\item \code{datasets}: List of datasets
}
}

\section{Methods}{

\itemize{
\item \code{cross_cor(key1=1, key2=2, method='pearson')}: Computes cross-dataset
correlation matrix between rows in two specified datasets.
\item \code{plot_cross_cor_heatmap(key1=1, key2=2, method='pearson', interactive=TRUE)}:
Plots multidataset correlation heatmap.
\item \code{print()}: Prints an overview of the object instance.
}
}

\section{Examples}{
\preformatted{TODO
}
}

