% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_factor_results}
\alias{get_factor_results}
\alias{get_numeric_results}
\title{Create EDA data for variables in a dataset}
\usage{
get_factor_results(df, target_var, ignore_vars = c(), type = "average")

get_numeric_results(df, target_var, ignore_vars = c(),
  type = "average", bins = 30)
}
\arguments{
\item{df}{The input \code{data.frame}}

\item{target_var}{A string with the name of the target variable}

\item{ignore_vars}{A character vector with the variables to ignore}

\item{type}{Can be \code{'count'} or \code{'average'} depicting how to get
the results back}

\item{bins}{Number of bins to use for numeric variables}
}
\description{
Get the EDA output with all the results for the factor or numeric variables
that are in the dataset \code{df}
}
