% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{get_factor_vars}
\alias{get_factor_vars}
\alias{get_numeric_vars}
\title{Get factor/numeric variables in a \code{data.frame}}
\usage{
get_factor_vars(df, target_var, ignore_vars = c())

get_numeric_vars(df, target_var, ignore_vars = c())
}
\arguments{
\item{df}{The input \code{data.frame}}

\item{target_var}{A string with the name of the target variable}

\item{ignore_vars}{A character vector with the variables to ignore}
}
\description{
Gets a character vector with all the factor/numeric variables in the specified
\code{data.frame} df.
}
