% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequency.R
\name{get_frequency}
\alias{get_frequency}
\title{get_frequency}
\usage{
get_frequency(df, vars, unique.threshold.pct = 0.8)
}
\arguments{
\item{df}{A dataframe of variables}

\item{vars}{A list of non-numeric variables to get frequencies for}

\item{unique.threshold.pct}{A value between 0 and 1. If the number of unique values for a variable is > to this threshold, it will not output the frequency distribution.}
}
\value{
A list of two.  Each being a dataframe.  First dataframe will give the frequency distribution to the variables.  The second dataframe will show the percent missing for each variable
}
\description{
Get frequency distribution of non-numeric variables
}
