% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numeric_bins.R
\name{get_numeric_bins}
\alias{get_numeric_bins}
\title{get_numeric_bins}
\usage{
get_numeric_bins(
  df,
  dv,
  dv.type,
  dv.denominator = NULL,
  var.list,
  nbins = 200,
  min.Pct = 0.05,
  binning.Type = "Bucketing",
  monotonic = TRUE,
  tracking = TRUE
)
}
\arguments{
\item{df}{A dataframe you are wanting to analyze}

\item{dv}{The name of the dependent variable (dv).  Example: 'target'}

\item{dv.type}{Can take on 1 of two inpunts - c('Binary','Frequency').  Both should be numeric.  If 'Frequency' is the input, it should be the numerator (if it is a rate).  The denominator will be specified as a separate parameter}

\item{dv.denominator}{The denominator of your dependent variable.  In many cases, this can be considered the exposure}

\item{var.list}{A list of non-numeric variables to analyze and create bins for}

\item{nbins}{Maximum number of bins to initially split the variable into}

\item{min.Pct}{The minimun percent of records a final bin should have.  Generally applies to only bins that are not NA}

\item{binning.Type}{The type of binning to use when splitting the variable.  One of two can be selected: c("Bucketing","Quantiles")}

\item{monotonic}{Logical TRUE/FALSE input.  If TRUE, it will force the bins to be monotonic based on the event rate}

\item{tracking}{Logical TRUE/FALSE input.  If set to TRUE, the user will be able to see what variable the function is analyzing.}
}
\value{
A list of dataframes.  First in the list will be 'Numeric_eda' - this is an aggregated dataframe showing the groups created along with other key information.  The second is 'numeric_iv' - This is a dataframe with each variable processed and their information value.  The last is 'numeric_logics' - This is a dataframe with the information needed to apply to your dataframe and transform your variables.  This table will be the input to apply_numeric_logic(logic_df=numeric_logics)
}
\description{
Numeric grouping
}
