% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentile.R
\name{get_percentile}
\alias{get_percentile}
\title{get_percentile}
\usage{
get_percentile(df, vars)
}
\arguments{
\item{df}{A dataframe of variables}

\item{vars}{A list of numeric variables to get percentile distributions for}
}
\value{
A dataframe with each numeric variable and the percentile distributions
}
\description{
Get percentile distribution of all numeric variables
}
