% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure.R
\name{get_structure}
\alias{get_structure}
\title{get_structure}
\usage{
get_structure(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe giving the structure of the input df
}
\description{
Get the structure of a dataframe
}
