% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_target_line}
\alias{get_target_line}
\title{Get the target variable line}
\usage{
get_target_line(df, target_var)
}
\arguments{
\item{df}{The input \code{data.frame}}

\item{target_var}{A string with the name of the target variable}
}
\description{
Simple function which gets the mean of the specified target variable in the
\code{data.frame} df. This is used to overlay on EDA plots.
}
