% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_buttons_module.R
\name{header_buttons}
\alias{header_buttons}
\title{Header Buttons Server Module}
\usage{
header_buttons(input, output, session, help_path = "docs/help.Rmd")
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{help_path}{path to help .Rmd file}
}
\value{
server
}
\description{
Header Buttons Server Module
}
