% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_buttons_module.R
\name{header_buttons_ui}
\alias{header_buttons_ui}
\title{Header Buttons UI Module}
\usage{
header_buttons_ui(id, contacts = NULL)
}
\arguments{
\item{id}{namespace ID}

\item{contacts}{contacts for contact dropdown}
}
\value{
tagList
}
\description{
Header Buttons UI Module
}
