% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{insertLogo}
\alias{insertLogo}
\title{Insert Logo}
\usage{
insertLogo(
  file,
  style = "background-color: #FFF; width: 100\%; height: 100\%;",
  width = NULL,
  ref = "#"
)
}
\arguments{
\item{file}{file}

\item{style}{style}

\item{width}{width}

\item{ref}{ref}
}
\value{
tag
}
\description{
Insert Logo
}
