% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owEDA-package.R
\docType{package}
\name{owEDA-package}
\alias{owEDA}
\alias{owEDA-package}
\title{owEDA: Oliver Wyman Exploratory Data Analysis Package}
\description{
Interactive EDA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig2011/owEDA}
  \item Report bugs at \url{https://github.com/jimbrig2011/owEDA/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

Other contributors:
\itemize{
  \item Scott Sobel \email{scott.sobel@oliverwyman.com} [contributor]
  \item Oliver Wyman Actuarial Consulting, Inc. [funder]
}

}
\keyword{internal}
