% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_all_analysis_results_vl}
\alias{plot_all_analysis_results_vl}
\title{Plot all EDA tables using Vega-Lite}
\usage{
plot_all_analysis_results_vl(analysis_results, which = "both",
  color_avg = "salmon", color_count = "skyblue")
}
\arguments{
\item{analysis_results}{The EDA analysis results as a \code{'data.frame'}}

\item{which}{Which plots to show, can be \code{'count'} plotting the
histograms, \code{'average'} plotting the averages of the target variable,
or \code{'both'} plotting both averages and histograms.}

\item{color_avg, color_count}{The fill color of the bars for averages and
counts.}
}
\description{
Create plots with all variables in the EDA table created using
\code{get_factor_results} or \code{get_numeric_results}.
}
