% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEDA.R
\name{plotEDA}
\alias{plotEDA}
\title{Plots Exploratory Data Analysis}
\usage{
plotEDA(df, variable = NULL, export = NULL, exportAll = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{variable}{A grouping variable.}

\item{export}{Option to export the results to a PDF.}

\item{exportAll}{Option to export plots by grouping variable to PDF.}
}
\value{
NULL.
}
\description{
This function receives as arguments a dataframe, a desired grouping variable
and an export option. It performs the graphical summary of the dataframe.
Depending on the argument variable, the output will be a summary of all the
variables of the dataframe (if 'variable' is null) or a summary focused
on one target variable. The type of the plots will change depending on
the 'variable' argument. The results of the function can be exported to a
PDF file, which name has to be specified when calling the function
(i.e., "Results", without ".pdf"). It does not plot variables if their number of
levels is higher than 10 (for categoric variables) or lower/equal to two
(for numeric variables). If 'variable' is not null but the exporting of plots to
a file is desired, changing the argument 'exportAll' to TRUE will generate
a PDF file for each grouping variable with the name of the original export
and the variable (i.e, "Results_variable.pdf").
}
