% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsByGroup.R
\name{statsByGroup}
\alias{statsByGroup}
\title{Statistics by Grouping Variable}
\usage{
statsByGroup(df, groupV, var, statistics = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{groupV}{A grouping variable.}

\item{var}{Dataframe variable in which metrics we are interested.}

\item{statistics}{Columns with p-value in numerical summaries.}
}
\value{
Grouped statistics by variable.
}
\description{
This function receives as arguments a dataframe, a desired grouping variable
and a variable. It returns a summary of the main statistic metrics of the
desired variable for each level of the grouping variable. Depending on the
levels of the grouping variable and the argument statistics, it will generate
a table with or without p-values (Student's t test and Mann-Whitney
non-parametric test)
}
