% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryByVariable.R
\name{summaryByVariable}
\alias{summaryByVariable}
\title{Summary by Variable}
\usage{
summaryByVariable(
  df,
  variable,
  export = NULL,
  exportAll = FALSE,
  statistics = TRUE
)
}
\arguments{
\item{df}{A dataframe.}

\item{variable}{A grouping variable.}

\item{export}{Option to export first exploratory plots to PDF.}

\item{exportAll}{Option to export plots by grouping variables to PDF.}

\item{statistics}{Addition of columns with p-Values for the levels of the variable.}
}
\value{
NULL.
}
\description{
This function receives as arguments a dataframe and a desired grouping variable.
The function will obtain the numeric and categoric variables of the dataframe
and will summon the statsByGroup function to create a table with the summary
numeric variables based on the grouping variable given as an argument and the
value of the statistics argument. Then, it will use Kable to correctly format
the table. Finally, it will summon the plotEDA function to generate the
graphic summary, also providing the desired arguments regarding the export
of plots to PDF files.
}
