% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data_module.R
\name{upload_data}
\alias{upload_data}
\title{Upload Data Module - Server}
\usage{
upload_data(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
list of uploaded files and data
}
\description{
Upload Data Module - Server
}
\examples{
\dontrun{
shiny::callModule(upload_data, "data")
}
}
