% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data_module.R
\name{upload_data_ui}
\alias{upload_data_ui}
\title{Upload Data Module - UI}
\usage{
upload_data_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
tagList of shinyFiles buttons
}
\description{
Upload Data Module - UI
}
\examples{
\dontrun{
upload_data_ui("data")
}
}
