% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractMultiDataSet.R
\name{AbstractMultiDataSet}
\alias{AbstractMultiDataSet}
\title{An R6 class representing a generic collection of datasets linked by
either column or row identifiers.}
\description{
Includes fucntionality and methods that are relevant to all EDA child
subclasses.
}
