% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioExprSet.R
\docType{data}
\name{BioExprSet}
\alias{BioExprSet}
\alias{BioExprSet}
\title{An S6 class representing an Expression dataset}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BioExprSet
}
\arguments{
\item{color}{Column metadata field to use for coloring points.}

\item{title}{Title to use for plot.}

\item{color}{Column metadata field to use for coloring points.}

\item{title}{Title to use for plot.}
}
\value{
A CPM-transformed version of the BioExprSet instance.
Log2 transforms data (adding 1 to ensure finite results).

Log2-transformed version of the expression data.
Plots a histogram of sample library sizes.

Generates a histogram of sample library sizes (sum of expression
levels within each column/sample).

A ggplot instance.
Plots bar graph of sample library sizes

Generates a bargraph plot of sample library sizes (sum of expression
levels within each column/sample). Each sample is shown as a separate
bar in the plot.

A ggplot instance.
Prints an overview of the object instance
Verifies that input data is an acceptable format.
}
\description{
BioExprSet is a simple class for interacting with biological expression
data (e.g. from microarray or RNA-Seq experiments.) It can accept either
a matrix and optional column/row metadata dataframes, or a Bioconductor
ExpressionSet instance. Methods are provided for common transformations
and visualizations.

Performs a counts-per-million (CPM) transformation.
}
\section{Arguments}{

\describe{
\item{dat}{matrix|ExpressionSet An expression dataset with rows corresponding
to genes, transcripts, probes, etc. and column corresponding to
individual samples.}
\item{row_mdata}{Data frame with rows corresponding to the column names of
\code{dat}.}
\item{row_mdata}{Data frame with rows corresponding to the row names of
\code{dat}.}
}
}

\keyword{datasets}
