% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{CarbonEmission}
\alias{CarbonEmission}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
Yongze Song \email{yongze.song@postgrad.curtin.edu.au}
}
\keyword{Carbon}
\keyword{Emission}
