% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_classes.R
\name{count_classes}
\alias{count_classes}
\title{Count class observations
Creates a new data frame with two columns,
listing the classes present in the input data frame,
and the number of observations for each class.}
\usage{
count_classes(data_frame, class_col)
}
\arguments{
\item{data_frame}{A data frame or data frame extension (e.g. a tibble).}

\item{class_col}{Unquoted column name of column containing class labels.}
}
\value{
A data frame with two columns.
The first column (named class) lists the classes from the input data frame.
The second column (named count) lists the number of observations for each class from the input data frame.
It will have one row for each class present in input data frame.
}
\description{
Count class observations
Creates a new data frame with two columns,
listing the classes present in the input data frame,
and the number of observations for each class.
}
\examples{
count_classes(mtcars, cyl)
}
