% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA.R
\name{EDA}
\alias{EDA}
\title{Exploratory Data Analysis (EDA)}
\usage{
EDA(
  df,
  rmCol = NULL,
  groupingVar = NULL,
  janitor = TRUE,
  fullTable = TRUE,
  statistics = TRUE,
  transf = NULL,
  export = NULL,
  exportAll = FALSE
)
}
\arguments{
\item{df}{A dataframe.}

\item{rmCol}{Vector of columns to remove of the dataframe.}

\item{groupingVar}{Vector of variables to perform summaries.}

\item{janitor}{Boolean. R-Friendly adaptation of variables and their levels.}

\item{fullTable}{Option to display a table with all variables (num and cat) or two tables (num and cat)}

\item{statistics}{p-Value columns in numerical summaries by grouping variable.}

\item{transf}{NatLog Transform of a numeric variable (or vector of numeric variables)}

\item{export}{Option to export the results to a PDF. Only name, without ".pdf".}

\item{exportAll}{Option to export plots by grouping variable to PDF.}
}
\value{
NULL.
}
\description{
This function receives several arguments: a dataframe, a vector with the
name of the uninformative columns that we want to remove, a vector with
the name of the desired grouping variables, a boolean to clean the variables
names and their levels with Janitor and an option to export the results
to a PDF file. It is the main function of the package and it will summon
basicSummary and summaryByVariable. Mainly, it will generate tables
for categoric and numeric variables (together or one table for each,
depending on the argument fullTable) with the desired metrics and its output
will be a formatted table with all these data. Then, it will summon the
summaryByVariable to obtain the plots and summaries for each grouping
variables (if needed). There are also additional options as to show
p-values with 2-levels grouping variables (argument statistics) and it
can also perform transformation of a variable of vector of variables to
natural logarithm (argument transf).
}
