% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDADataFrame.R
\docType{data}
\name{EDADataFrame}
\alias{EDADataFrame}
\alias{EDADataFrame}
\title{An S6 class representing a dataframe dataset.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EDADataFrame
}
\description{
EDADataFrame is a helper class for wrapping dataframes, with optional
support for row and column datadata. Methods are provided for common
exploratory data analysis summary statistics, transformations, and
visualizations.

EDADataFrame constructor
Prints an overview of the object instance
Verifies that input data is of type dataframe
}
\section{Arguments}{

\describe{
\item{dat}{An m x n dataset.}
\item{row_mdata}{Data frame with rows corresponding to the column names of
\code{dat}.}
\item{row_mdata}{Data frame with rows corresponding to the row names of
\code{dat}.}
}
}

\keyword{datasets}
