% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDADataSet.R
\docType{data}
\name{EDADataSet}
\alias{EDADataSet}
\alias{EDADataSet}
\title{An S6 class representing a generic dataset.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EDADataSet
}
\value{
None
}
\description{
An S6 class representing a generic dataset.

Based on code adapted from cbcbSEQ
(https://github.com/kokrah/cbcbSEQ/) originally written by Kwame Okrah.
}
\section{Arguments}{

\describe{
  \item{dat}{An m x n dataset.}
  \item{row_mdata}{Data frame with rows corresponding to the row names of 
      \code{dat}.}
  \item{row_mdata}{Data frame with rows corresponding to the column names of 
      \code{dat}.}
}
}

\author{
Kwame Okrah

V. Keith Hughitt, \email{khughitt@umd.edu}

Plot median pairwise variable correlations

Plots the median correlation of each variable (column)

V. Keith Hughitt, \email{keith.hughitt@nih.gov}
}
\keyword{datasets}
