% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDADataSet.R
\docType{data}
\name{EDADataSet}
\alias{EDADataSet}
\alias{EDADataSet}
\title{An S6 class representing a generic dataset.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EDADataSet
}
\arguments{
\item{mask}{Logical vector of length equal to the number of rows in
the dataset.}

\item{mask}{Logical vector of length equal to the number of columns
in the dataset.}

\item{method}{Character array specifying imputation method to use
(knn)
Subsample dataset

Randomly subsamples dataset and returns a new EDADataSet instance.
For removing specific rows and columns, see the \code{filter_rows} and
\code{filter_cols} functions.
Summarizes overall characteristics of a dataset

Kernel density plot

Plots densities for each row or column in the dataset. This is most
useful when you are interested in similarties or differences in
distributions across columns, for a relatively small number of
columns.}

\item{color}{Variable to color density curves by. If not is
specified, uses variable specified at object construction time,
or else, uses a separate color for each column.}

\item{color}{Color variable as passed into plot function call}

\item{color}{Color variable as passed into plot function call}

\item{color}{Color variable as passed into plot function call}

\item{shape}{Shape variable as passed into plot function call}

\item{styles}{List of color-related style info}

\item{color}{Color variable passed into plot function call
Determines shape-related style information to use for a plot}

\item{styles}{List of shape-related style info}

\item{shape}{shape variable passed into plot function call
Returns a list of color-related style information}

\item{color}{Color variable as passed into plot function call}

\item{shape}{Shape variable as passed into plot function call}

\item{color}{Variable to use for assigning colors.}

\item{color_pal}{Color palette to use}

\item{shape}{Variable to use for assigning shapes.}

\item{dat}{Dataset}

\item{row_ids}{Column id or number where row identifiers are stored.}

\item{col_ids}{Row id or number where column identifiers are stored.}

\item{Dataset}{with identifiers as rows and columns
Prints dataset summary to screen
Determines metadata columns to include for heatmap, etc. functions}

\item{include}{Vector of strings indicating metadata columns which
should be included in the analysis.}

\item{exclude}{Vector of strings indicating metadata columns which
should be excluded from the analysis.}
}
\value{
A filtered version of the original EDADataSet object.
Applies a filter to columns of the dataset

A filtered version of the original EDADataSet object.
Imputes missing values in the dataset

Imputes missing values in the dataset using a specified method
and stores the result in-place. Currently only support k-Nearest
Neighbors (kNN) method.

Note: When using the \code{knn} method, it may be neccessary to set R
the environmental variable \code{R_MAX_NUM_DLLS} to some value larger
than its default of 100 (150 should be sufficient), prior to
launching R. This can be set in the ~/.Renviron.

ggplot plot instance.

Prints class greeting to the screen
Generates ggplot aesthetics for density plots

List of style information
Generates ggplot aesthetics for histogram plots

List of style information
Generates ggplot aesthetics for a geom_point plot

List of geom_point style information
Determines color-related style information to use for a plot

list List of color-related properties
Returns a list of shape-related style information

list List of shape-related properties
Returns a vector of color codes associated with the specified
variable.

Vector of colors with length equal to the number of columns
in the data.
Returns a vector of text labels to use for a plot

Vector of labels with length equal to the number of columns
in the data.
Normalizes handling of data row and column identifiers

Checks dataset row and column identifiers and converts them to row
and column names, respectively, if they are not already stored there.

Character vector of fields to include in analysis.
Transpose the dataset and metadata in-place
}
\description{
An S6 class representing a generic dataset.

Clears any cached resuts and performs garbage collection to free
up memory.
Applies a filter to rows of the dataset
}
\section{Arguments}{

\describe{
\item{dat}{An m x n dataset.}
\item{row_mdata}{Data frame with rows corresponding to the column names of
\code{dat}.}
\item{row_mdata}{Data frame with rows corresponding to the row names of
\code{dat}.}
}
}

\keyword{datasets}
