% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDADataSet.R
\name{EDADataSet}
\alias{EDADataSet}
\title{An S6 class representing a generic dataset.}
\description{
EDADataSet is a helper class for wrapping data matrices and data frames,
with optional support for row and column datadata. Methods are provided for
common exploratory data analysis summary statistics, transformations, and
visualizations.
}
\details{
Rather than using EDADataSet directly, it is instead recommended that you
use one of the more specific child classes:
\itemize{
\item \code{EDADataFrame}
\item \code{EDAMatrix}
\item etc.
}

The child classes inherit from \code{EDADataSet}, and as such, will include all
of the same functionality, but with some additional support for the specific
data type they are designed for.
}
