% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAMatrix.R
\name{EDAMatrix}
\alias{EDAMatrix}
\title{An S6 class representing a matrix dataset.}
\description{
An S6 class representing a matrix dataset.
}
\section{Arguments}{

\describe{
  \item{dat}{An m x n dataset.}
  \item{row_mdata}{Data frame with rows corresponding to the row names of 
      \code{dat}.}
  \item{row_mdata}{Data frame with rows corresponding to the column names of 
      \code{dat}.}
}
}

