% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAMultiDataSet.R
\name{EDAMultiDataSet}
\alias{EDAMultiDataSet}
\title{An S6 class representing collection of related datasets}
\description{
An S6 class representing collection of related datasets
}
\section{Usage}{
\preformatted{# TODO
}
}

\section{Arguments}{

\itemize{
\item \code{...}: Two or more EDADataSet instances, indexed by the same column
identifiers.
}
}

\section{Fields}{

\itemize{
\item \code{datasets}: List of datasets
}
}

\section{Methods}{

\itemize{
\item \code{cross_cor(key1=1, key2=2, method='pearson')}: Computes cross-dataset
correlation matrix between rows in two specified datasets.
\item \code{plot_cross_cor_heatmap(key1=1, key2=2, method='pearson', interactive=TRUE)}:
Plots multidataset correlation heatmap.
\item \code{print()}: Prints an overview of the object instance.
}
}

\section{Examples}{
\preformatted{TODO
}
}

