% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMDI.R
\name{LMDI}
\alias{LMDI}
\title{Log Mean Devisia Index method for energy decomposition analysis}
\usage{
LMDI(C0, CT, X0, XT)
}
\arguments{
\item{C0}{A data.frame}

\item{CT}{A data.frame}

\item{X0}{A data.frame}

\item{XT}{A data.frame}
}
\description{
Log Mean Devisia Index method for energy decomposition analysis
}
\examples{
data(CarbonEmission)
carbon <- CarbonEmission$carbon
cdata <- carbon[,-c(1,2)]
xdata <- CarbonEmission$xdata
C0 <- cdata[1,]
CT <- cdata[2,]
X0 <- xdata[[1]][1,]
XT <- xdata[[2]][2,]
LMDI(C0, CT, X0, XT)

}
