% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owEDA-package.R
\docType{package}
\name{owEDA-package}
\alias{owEDA}
\alias{owEDA-package}
\title{owEDA: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

}
\keyword{internal}
