% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{plot_analysis_category}
\alias{plot_analysis_category}
\title{Plot EDA tables}
\usage{
plot_analysis_category(analysis_result, type = "count",
  position = "fill")
}
\arguments{
\item{analysis_result}{The EDA analysis result as a \code{'data.frame'}}

\item{type}{Can be \code{'count'} to plot the EDA resuls when calculated
using counts, \code{'average'} to plot the results when calculated using
the average of the target variable or \code{'count_all'} to get a count of
all the values in the variables}
}
\description{
Create plots with all variables in the EDA table created using
\code{create_analysis_table} or \code{create_numeric_table}.
}
